+++ src/viam/sdk/components/motor/client.cpp
@@ -1,4 +1,3 @@
-#include "viam/sdk/common/proto_type.hpp"
 #include <viam/sdk/components/motor/client.hpp>
 
 #include <algorithm>
@@ -10,7 +9,7 @@
 #include <viam/api/common/v1/common.pb.h>
 #include <viam/api/component/motor/v1/motor.grpc.pb.h>
 
-#include <viam/sdk/common/utils.hpp>
+#include <viam/sdk/common/client_helper.hpp>
 #include <viam/sdk/components/motor/motor.hpp>
 #include <viam/sdk/config/resource.hpp>
 #include <viam/sdk/robot/client.hpp>
@@ -22,77 +21,6 @@ MotorClient::MotorClient(std::string name, std::shared_ptr<grpc::Channel> channe
     : Motor(std::move(name)),
       stub_(viam::component::motor::v1::MotorService::NewStub(channel)),
       channel_(std::move(channel)){};
-
-template <typename ClientType, typename StubType, typename RequestType, typename ResponseType>
-class ClientHelper {
-    static void default_rsc_(RequestType&) {}
-    static void default_rhc_(const ResponseType&) {}
-    static void default_ehc_(const ::grpc::Status& status) {
-        throw std::runtime_error(status.error_message());
-    }
-
-   public:
-    using PFn = ::grpc::Status (StubType::*)(::grpc::ClientContext*,
-                                             const RequestType&,
-                                             ResponseType*);
-    explicit ClientHelper(ClientType* client, StubType* stub, PFn pfn)
-        : client_(client), stub_(stub), pfn_(pfn) {}
-
-    ClientHelper& with(const AttributeMap& extra) {
-        return with(extra, default_rsc_);
-    }
-
-    template <typename RequestSetupCallable>
-    ClientHelper& with(RequestSetupCallable&& rsc) {
-        std::forward<RequestSetupCallable>(rsc)(request_);
-        return *this;
-    }
-
-    template <typename RequestSetupCallable>
-    ClientHelper& with(const AttributeMap& extra, RequestSetupCallable&& rsc) {
-        *request_.mutable_extra() = map_to_struct(extra);
-        return with(std::forward<RequestSetupCallable>(rsc));
-    }
-
-    template <typename ResponseHandlerCallable = decltype(default_rhc_)>
-    auto invoke(ResponseHandlerCallable&& rhc = default_rhc_) {
-        return invoke(std::forward<ResponseHandlerCallable>(rhc), default_ehc_);
-    }
-
-    template <typename ResponseHandlerCallable, typename ErrorHandlerCallable>
-    auto invoke(ResponseHandlerCallable&& rhc, ErrorHandlerCallable&& ehc) {
-        *request_.mutable_name() = client_->name();
-        ::grpc::ClientContext ctx;
-        set_client_ctx_authority(ctx);
-
-        const auto result = (stub_->*pfn_)(&ctx, request_, &response_);
-        if (result.ok()) {
-            return std::forward<ResponseHandlerCallable>(rhc)(
-                const_cast<const ResponseType&>(response_));
-        }
-
-        std::forward<ErrorHandlerCallable>(ehc)(result);
-        std::abort();
-    }
-
-   private:
-    ClientType* client_;
-    StubType* stub_;
-    PFn pfn_;
-    RequestType request_;
-    ResponseType response_;
-};
-
-// TODO: Stop returns the status???
-
-template <typename ClientType, typename StubType, typename RequestType, typename ResponseType>
-auto make_client_helper(ClientType* client,
-                        StubType& stub,
-                        ::grpc::Status (StubType::*method)(::grpc::ClientContext*,
-                                                           const RequestType&,
-                                                           ResponseType*)) {
-    return ClientHelper<ClientType, StubType, RequestType, ResponseType>(client, &stub, method);
-}
 
 void MotorClient::set_power(double power_pct, const AttributeMap& extra) {
     return make_client_helper(this, *stub_, &StubType::SetPower)
